+++ src/Ext/TechnoType/Body.cpp
@@ -1208,6 +1208,8 @@ void TechnoTypeExt::LoadFromINIFile(CCINIClass* const pINI)
 
 	exINI.ReadSpeed(pSection, "SubterraneanSpeed", &this->SubterraneanSpeed);
 	this->SubterraneanHeight.Read(exINI, pSection, "SubterraneanHeight");
+	
+	this->VoiceEnterGrinder.Read(exINI, pSection, "VoiceEnterGrinder");
 
 	// Ares 0.2
 	this->RadarJamRadius.Read(exINI, pSection, "RadarJamRadius");
@@ -1847,6 +1849,8 @@ void TechnoTypeExt::Serialize(T& Stm)
 
 		.Process(this->SubterraneanSpeed)
 		.Process(this->SubterraneanHeight)
+
+		.Process(this->VoiceEnterGrinder)
 
 		// Ares 0.2
 		.Process(this->RadarJamRadius)